Fixes #172 - Hide Call for Proposals page from UI
