+++ client.jsx
@@ -6,6 +6,7 @@ var routes = (
   <Route>
     <Route name="home" path="/" handler={require('./pages/home.jsx')} />
     <Route name="proposals" path="/proposals" handler={require('./pages/proposals.jsx')} />
+    <Route name="proposals-extended" path="/proposals-2015" handler={require('./pages/proposals-extended.jsx')} />
     <Route name="tickets" path="/tickets" handler={require('./pages/tickets.jsx')} />
     <Route name="location" path="/location" handler={require('./pages/location.jsx')} />
     <Route name="about" path="/about" handler={require('./pages/about.jsx')} />
