+++ less/pages/proposals.less
@@ -1,18 +1,6 @@
 .proposals-page {
   .hero-unit {
     background-position: center 14%;
-    .deadline {
-      margin-top: 5px;
-      span {
-        font-size: 20px;
-        text-transform: none;
-        font-weight: 400;
-        letter-spacing: normal;
-        color: #FFF;
-        background: rgba(93, 160, 143, 0.85);
-        padding: 12px 19px;
-      }
-    }
   }
   .hero-unit-container {
     background: linear-gradient( to right, #599D73, #57949D);
